Extract dequeue_coordinator into its own file
